+++ history.md
@@ -16,7 +16,6 @@
 
 ## High priority features planned
 - []  (feature) Creating thumbnails from Web Interface (including status)
-- []  (bugfix) When import a non supported image Ok is shown 
 - [] Change from `Newtonsoft.Json` to `System.Text.Json` __not implemented__
 
 # version 0.1.11 - 2019-12-??
@@ -32,6 +31,7 @@
 - [x]   (bugfix) _Front-end_ change default outline for Chrome/Safari
 - [x]   (bugfix) _Front-end_ fix document.title undefined error 
 - [x]   (bugfix) _Front-end_ fix issue where on empty searchquery a sidebar is shown
+- [x]   (bugfix) _Front-end_ When import a non supported image Ok is shown 
 
 # version 0.1.10 - 2019-12-15
 - [x] (bugfix) Archive => After pressing 'Apply' the updates are not shown